Use common stats class for duplicated code
