+++ server/routes/key-worker-statistics/controller.ts
@@ -3,6 +3,7 @@ import KeyworkerApiService from '../../services/keyworkerApi/keyworkerApiService
 import { components } from '../../@types/keyWorker'
 import { formatDateConcise } from '../../utils/datetimeUtils'
 import { FLASH_KEY__FORM_RESPONSES } from '../../utils/constants'
+import { formatChange, formatNumber, formatValue } from '../../utils/statsUtils'
 
 export class KeyWorkerStatisticsController {
   constructor(private readonly keyworkerApiService: KeyworkerApiService) {}
@@ -105,24 +106,4 @@ export class KeyWorkerStatisticsController {
     )
     return res.redirect('/key-worker-statistics')
   }
-}
-
-function formatNumber(value: string | number, type?: string) {
-  if (type === 'percentage') return Number(parseFloat(value.toString()).toFixed(2))
-  return Number(value)
-}
-
-function formatValue(value: number, type: string) {
-  if (type === 'percentage') return `${value} %`
-  if (value === 0) return 0
-  if (!value) return '-'
-  return Number(value)
-}
-
-function formatChange(change: number, type: string) {
-  const formatted = formatValue(change, type)
-  if (change > 0) return `+${formatted} increase`
-  if (change < 0) return `${formatted} decrease`
-  if (change === 0) return 'No change'
-  return formatted
 }
